Try to fix circular dependency
